remove symbolic link from settings file if it cannot be found.
