+++ XRTK-Core/Assets/XRTK/Inspectors/Utilities/SymbolicLinks/SymbolicLinkSettingsInspector.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 31f612492a45459bb8690552696cab55
+timeCreated: 1554672145
